Add variable ssh_parent_wait_sleep for specifying wait extra time for child SSH process
This is necessary in the case where the child SSH process takes longer
than 3 seconds to open a local port for forwarding.

I notice this when I try to use terraform-ssh-tunnel with a host that is
on the other side of the world from me.  It takes longer than 3 seconds
for the child SSH process to actually open the local port that is being
forwarded.

Without a longer wait time, I was seeing errors in Terraform when trying
to use the local port.